requirements.txt
streamlit
pyyaml
rdkit-pypi
